requirements.txt
requests-sse>=0.3.2
python-dotenv==1.0.1
confluent-kafka==2.8.0